dev: move the location of home
