+++ iwfpapp/test/integration/nav_tab_test.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:iwfpapp/services/data_store.dart';
 import 'package:iwfpapp/services/mode.dart';
-import 'package:iwfpapp/nav.dart';
-import './mock_mode.dart';
-import './mock_data_store.dart';
+import 'package:iwfpapp/screens/home/main.dart';
+import 'mock_mode.dart';
+import 'mock_data_store.dart';
 
 void main() {
   testWidgets('test app render no crash', (WidgetTester tester) async {
